feat: soft delete (#622)
## What type of PR is this?

/kind feature

## What this PR does / why we need it:

In the current implementation, the lifecycle of storage data is bound to
object in cluster. When an object is deleted in cluster, the associated
data in storage will also be cleared.

Events in cluster are retained for one hour by default. As key
information for troubleshooting, it is usually hoped that events can be
retained for a longer time, which requires karpor to unbind the
lifecycle of specific types of resources from cluster.

This PR adds a `remainAfterDeleted` field in the SyncRegistry CRD to
control that a certain type of resources remain in storage after being
deleted in cluster.

If the user sets `remainAfterDeleted=true` for a type of resource, the
user needs to delete the associated data in the storage by himself.
Karpor only marks `deleted=true` for storage data when the object is
deleted in cluster.

For resources that leave `remainAfterDeleted` unset
(`remainAfterDeleted=false`), the data in storage will be deleted
immediately after the cluster object is deleted.

```yaml
apiVersion: search.karpor.io/v1beta1
kind: SyncRegistry
metadata:
  name: events-sync-rule
spec:
  clusters:
    - "*"
  syncResources:
    - apiVersion: v1
      resource: events
      trimRefName: for-events
      remainAfterDeleted: true
```

## Which issue(s) this PR fixes:

<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->

Fixes #

---------

Co-authored-by: tianahi <wb-th358723@antgroup.com>