docs: update Slack channels in README (#647)
## What type of PR is this?

/kind documentation

## What this PR does / why we need it:

This PR updates the Slack channel links in the README files to point to
the cloud-native workspace. It also adds descriptions to distinguish
between the #kusion and #kusion-general channels. The changes are
applied to all README versions (English, Chinese, Portuguese) and the
formatting is improved with clear channel section headers.

## Which issue(s) this PR fixes:

Fixes #