refactor(ui): standardize event and log tab names (#667)
## What type of PR is this?

/kind refactor

## What this PR does / why we need it:

This PR standardizes the names of the 'Event' and 'Log' tabs across UI
components. Specifically, it renames 'Event' to 'Events' and 'Log' to
'Logs'. Additionally, it updates the i18n keys to match the new
component names and fixes tab value handling in insight detail pages.
This change introduces a breaking change where the tab values for Event
and Log have been renamed to Events and Logs respectively.

## Which issue(s) this PR fixes:

Fixes #